Upgrade subxt and the node metadata
