+++ crates/phaxt/src/lib.rs
@@ -1,4 +1,4 @@
-use parity_scale_codec::{Encode, Decode};
+use parity_scale_codec::{Decode, Encode};
 use scale_info::TypeInfo;
 
 #[subxt::subxt(runtime_metadata_path = "metadata_files/khala_metadata.scale")]
@@ -12,25 +12,37 @@ pub mod khala {
 #[subxt::subxt(runtime_metadata_path = "metadata_files/kusama_metadata.scale")]
 pub mod kusama {}
 
-#[derive(Encode, Decode, Clone, PartialEq, Eq, TypeInfo, PartialOrd, Ord)]
+#[derive(Encode, Decode, Clone, PartialEq, Eq, TypeInfo, PartialOrd, Ord, Debug)]
 pub struct ParaId(pub u32);
 
-pub mod extra;
 pub mod rpc;
-pub mod khala_config;
-pub use khala_config::*;
 
-pub type PhalaExtra = extra::PhalaExtra<KhalaConfig>;
-
-pub type ParachainApi = khala::RuntimeApi<KhalaConfig, PhalaExtra>;
-pub type RelaychainApi = kusama::RuntimeApi<DefaultConfig, DefaultExtra<DefaultConfig>>;
+pub type ParachainApi = khala::RuntimeApi<DefaultConfig, ExtrinsicParams>;
+pub type RelaychainApi = kusama::RuntimeApi<DefaultConfig, ExtrinsicParams>;
+pub type ExtrinsicParams = DefaultExtrinsicParams<DefaultConfig>;
 
 pub use subxt;
 pub use subxt::sp_core::storage::{StorageData, StorageKey};
 
-use subxt::{DefaultExtra, DefaultConfig};
+pub use subxt::{
+    extrinsic::{
+        PolkadotExtrinsicParams as DefaultExtrinsicParams,
+        PolkadotExtrinsicParamsBuilder as ExtrinsicParamsBuilder,
+    },
+    DefaultConfig,
+};
 
 pub fn storage_key<T: subxt::StorageEntry>(entry: T) -> StorageKey {
     let prefix = subxt::storage::StorageKeyPrefix::new::<T>();
     entry.key().final_key(prefix)
 }
+
+pub type Index = <DefaultConfig as subxt::Config>::Index;
+pub type BlockNumber = <DefaultConfig as subxt::Config>::BlockNumber;
+pub type Hash = <DefaultConfig as subxt::Config>::Hash;
+pub type Hashing = <DefaultConfig as subxt::Config>::Hashing;
+pub type AccountId = <DefaultConfig as subxt::Config>::AccountId;
+pub type Address = <DefaultConfig as subxt::Config>::Address;
+pub type Header = <DefaultConfig as subxt::Config>::Header;
+pub type Signature = <DefaultConfig as subxt::Config>::Signature;
+pub type Extrinsic = <DefaultConfig as subxt::Config>::Extrinsic;
